requirements.txt
# Core Dependencies
PyQt5>=5.15.0,<6.0.0
python-vlc>=3.0.0,<4.0.0

# Media Processing
mutagen>=1.45.0,<2.0.0
Pillow>=9.0.0,<10.0.0

# Development
black>=22.3.0
flake8>=4.0.0
isort>=5.10.0
pytest>=7.0.0
